Refactor author block classes; replace base class with Author_Block_Base, streamline error handling, and enhance caching mechanisms
