+++ includes/Blocks/Author_List_Block.php
@@ -9,7 +9,7 @@
 
 namespace AuthorProfileBlocks\Blocks;
 
-use AuthorProfileBlocks\Plugin;
+use AuthorProfileBlocks\Common\Author_Block_Base;
 use WP_Block;
 
 // Exit if accessed directly.
@@ -20,14 +20,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 /**
  * Class that handles the Author List block.
  */
-class Author_List_Block extends Block_Base {
-	/**
-	 * Cache for rendered author lists to avoid duplicate processing.
-	 *
-	 * @var array
-	 */
-	private array $list_cache = array();
-
+class Author_List_Block extends Author_Block_Base {
 	/**
 	 * Get the block name.
 	 *
@@ -35,41 +28,6 @@ class Author_List_Block extends Block_Base {
 	 */
 	public function get_block_name(): string {
 		return 'author-list';
-	}
-
-	/**
-	 * Additional initialization actions for the block.
-	 *
-	 * @return void
-	 */
-	protected function additional_init(): void {
-		add_filter( 'render_block_author-profile-blocks/author-list', array( $this, 'filter_rendered_output' ), 10, 2 );
-	}
-
-	/**
-	 * Filter the rendered output of the block.
-	 * This is a final processing step after the block renders.
-	 *
-	 * @param string $block_content The rendered block content.
-	 * @param array  $block         The block data.
-	 *
-	 * @return string The filtered content.
-	 */
-	public function filter_rendered_output( string $block_content, array $block ): string {
-		/**
-		 * Filters the rendered output of the author list block.
-		 *
-		 * This filter allows themes and plugins to modify the final HTML output
-		 * of the Author List block before it's displayed on the frontend.
-		 *
-		 * @since 1.0.0
-		 *
-		 * @param string $block_content The rendered HTML output of the block.
-		 * @param array  $block         The block attributes and content.
-		 *
-		 * @return string               The filtered block content.
-		 */
-		return apply_filters( 'author_profile_blocks_rendered_list', $block_content, $block );
 	}
 
 	/**
@@ -94,15 +52,14 @@ class Author_List_Block extends Block_Base {
 		$author_ids = $attributes['authorIds'] ?? array();
 
 		if ( empty( $author_ids ) ) {
-			return '<div class="apb-author-list-error">' .
-			       esc_html__( 'Please select authors for the list.', 'author-profile-blocks' ) .
-			       '</div>';
+			return $this->render_error_message( __( 'Please select authors for the list.', 'author-profile-blocks' ) );
 		}
 
 		// Check cache first.
 		$cache_key = $this->generate_cache_key( $author_ids, $attributes );
-		if ( isset( $this->list_cache[ $cache_key ] ) ) {
-			return $this->list_cache[ $cache_key ];
+		$cached_content = $this->get_cached_render( $cache_key );
+		if ( $cached_content ) {
+			return $cached_content;
 		}
 
 		// Apply author role filter if specified.
@@ -111,51 +68,30 @@ class Author_List_Block extends Block_Base {
 			$roles = array( $attributes['authorRole'] );
 		}
 
-		// Get authors data using Plugin instance with role filtering.
-		$authors = array();
-		$plugin  = Plugin::get_instance();
-
-		// Handle individual author IDs.
-		foreach ( $author_ids as $author_id ) {
-			$author_data = $plugin->get_author_data( $author_id );
-			if ( $author_data ) {
-				// Apply role filter if specified.
-				if ( ! empty( $roles ) && ! in_array( $author_data['role'], $roles, true ) ) {
-					continue;
-				}
-				$authors[] = $author_data;
-			}
-		}
+		// Get authors data.
+		$authors = $this->get_authors_data( $author_ids, $roles );
 
 		// Apply maximum authors limit if specified.
 		$max_authors = isset( $attributes['maxAuthors'] ) ? (int) $attributes['maxAuthors'] : 0;
-		if ( $max_authors > 0 && count( $authors ) > $max_authors ) {
-			$authors = array_slice( $authors, 0, $max_authors );
-		}
+		$authors = $this->apply_author_limit( $authors, $max_authors );
 
 		// If no authors found after filtering.
 		if ( empty( $authors ) ) {
-			return '<div class="apb-author-list-error">' .
-			       esc_html__( 'No authors found matching the specified criteria.', 'author-profile-blocks' ) .
-			       '</div>';
+			return $this->render_error_message( __( 'No authors found matching the specified criteria.', 'author-profile-blocks' ) );
 		}
 
 		// Generate styles for the block.
-		$wrapper_styles  = $this->get_block_styles( $attributes );
+		$wrapper_styles = $this->get_block_styles( $attributes );
 		$style_attribute = '';
 
 		if ( ! empty( $wrapper_styles ) ) {
-			$style_strings = array();
-			foreach ( $wrapper_styles as $property => $value ) {
-				$style_strings[] = $property . ': ' . $value;
-			}
-			$style_attribute = esc_attr( implode( '; ', $style_strings ) );
+			$style_attribute = $this->get_styles_string( $wrapper_styles );
 		}
 
 		// Classes for the block wrapper.
 		$wrapper_attributes = get_block_wrapper_attributes(
 			array(
-				'class' => $this->get_block_classes( $attributes ),
+				'class' => $this->get_block_classes( $attributes, 'list' ),
 				'style' => $style_attribute,
 			)
 		);
@@ -192,7 +128,7 @@ class Author_List_Block extends Block_Base {
 		$html .= '</div>'; // Close block wrapper.
 
 		// Cache the result.
-		$this->list_cache[ $cache_key ] = $html;
+		$this->set_cached_render( $cache_key, $html );
 
 		return $html;
 	}
@@ -207,15 +143,11 @@ class Author_List_Block extends Block_Base {
 	 */
 	private function render_author_item( array $author, array $attributes ): string {
 		// Get item styles.
-		$item_styles     = $this->get_item_styles( $attributes );
+		$item_styles = $this->get_item_styles( $attributes );
 		$style_attribute = '';
 
 		if ( ! empty( $item_styles ) ) {
-			$style_strings = array();
-			foreach ( $item_styles as $property => $value ) {
-				$style_strings[] = $property . ': ' . $value;
-			}
-			$style_attribute = ' style="' . esc_attr( implode( '; ', $style_strings ) ) . '"';
+			$style_attribute = ' style="' . $this->get_styles_string( $item_styles ) . '"';
 		}
 
 		// Item classes based on layout and options.
@@ -240,323 +172,14 @@ class Author_List_Block extends Block_Base {
 		// Layout depends on display style.
 		$display_style = $attributes['displayStyle'] ?? 'compact';
 
-		switch ( $display_style ) {
-			case 'detailed':
-				$html .= $this->render_detailed_layout( $author, $attributes );
-				break;
-
-			case 'compact':
-			default:
-				$html .= $this->render_compact_layout( $author, $attributes );
-				break;
+		if ( $display_style === 'detailed' ) {
+			$html .= $this->render_detailed_layout( $author, $attributes );
+		} else {
+			$html .= $this->render_compact_layout( $author, $attributes );
 		}
 
 		$html .= '</div>'; // Close .apb-author-list-item-content.
 		$html .= '</li>'; // Close list item.
-
-		return $html;
-	}
-
-	/**
-	 * Generate a cache key for the author list based on attributes.
-	 *
-	 * @param array $author_ids The author IDs.
-	 * @param array $attributes The block attributes.
-	 *
-	 * @return string The cache key.
-	 */
-	private function generate_cache_key( array $author_ids, array $attributes ): string {
-		// Sort author IDs to ensure consistent cache key regardless of order.
-		sort( $author_ids );
-
-		return md5( implode( ',', $author_ids ) . maybe_serialize( $attributes ) );
-	}
-
-	/**
-	 * Get block classes based on attributes.
-	 *
-	 * @param array $attributes Block attributes.
-	 *
-	 * @return string CSS classes.
-	 */
-	private function get_block_classes( array $attributes ): string {
-		$classes = array();
-
-		// Text alignment.
-		if ( ! empty( $attributes['textAlign'] ) ) {
-			$classes[] = 'has-text-align-' . $attributes['textAlign'];
-		}
-
-		// Display style.
-		$display_style = $attributes['displayStyle'] ?? 'compact';
-		$classes[]     = 'is-style-' . $display_style;
-
-		return implode( ' ', $classes );
-	}
-
-	/**
-	 * Get block styles based on attributes.
-	 *
-	 * @param array $attributes Block attributes.
-	 *
-	 * @return array CSS styles.
-	 */
-	private function get_block_styles( array $attributes ): array {
-		$styles = array();
-
-		// Background color.
-		if ( ! empty( $attributes['backgroundColor'] ) ) {
-			$styles['background-color'] = $attributes['backgroundColor'];
-		}
-
-		// Padding.
-		if ( isset( $attributes['blockPadding'] ) ) {
-			$styles['padding'] = $attributes['blockPadding'] . 'px';
-		}
-
-		return $styles;
-	}
-
-	/**
-	 * Get styles for individual list items.
-	 *
-	 * @param array $attributes Block attributes.
-	 *
-	 * @return array CSS styles.
-	 */
-	private function get_item_styles( array $attributes ): array {
-		$styles = array();
-
-		// Item background color.
-		if ( ! empty( $attributes['itemBackgroundColor'] ) ) {
-			$styles['background-color'] = $attributes['itemBackgroundColor'];
-		}
-
-		// Padding.
-		if ( isset( $attributes['itemPadding'] ) ) {
-			$styles['padding'] = $attributes['itemPadding'] . 'px';
-		}
-
-		// Border color if dividers enabled.
-		if ( ! empty( $attributes['enableDividers'] ) && ! empty( $attributes['dividerColor'] ) ) {
-			$styles['border-color'] = $attributes['dividerColor'];
-		}
-
-		return $styles;
-	}
-
-	/**
-	 * Render compact author list item layout.
-	 *
-	 * @param array $author     Author data.
-	 * @param array $attributes Block attributes.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_compact_layout( array $author, array $attributes ): string {
-		$html = '<div class="apb-author-list-compact">';
-
-		// Author image - only if image display is enabled in attributes.
-		if ( ! empty( $author['image'] ) && ( ! isset( $attributes['showImage'] ) || $attributes['showImage'] ) ) {
-			$html .= $this->render_author_image( $author );
-		}
-
-		// Author info container.
-		$html .= '<div class="apb-author-list-info">';
-
-		// Author name.
-		if ( ! empty( $author['title'] ) ) {
-			$html .= $this->render_author_name( $author );
-		}
-
-		// Author position if available and display is enabled.
-		if ( ! empty( $author['position'] ) && ( ! isset( $attributes['showPosition'] ) || $attributes['showPosition'] ) ) {
-			$html .= $this->render_author_position( $author );
-		}
-
-		$html .= '</div>'; // Close .apb-author-list-info.
-
-		// Social icons if enabled.
-		if ( ! empty( $author['social'] ) && is_array( $author['social'] ) && ( ! isset( $attributes['showSocial'] ) || $attributes['showSocial'] ) ) {
-			$html .= '<div class="apb-author-list-social">';
-			$html .= $this->render_social_profiles( $author['social'], 'apb-list-social' );
-			$html .= '</div>';
-		}
-
-		$html .= '</div>'; // Close .apb-author-list-compact.
-
-		return $html;
-	}
-
-	/**
-	 * Render detailed author list item layout.
-	 *
-	 * @param array $author     Author data.
-	 * @param array $attributes Block attributes.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_detailed_layout( array $author, array $attributes ): string {
-		$html = '<div class="apb-author-list-detailed">';
-
-		// Left column with image.
-		$html .= '<div class="apb-author-list-left">';
-
-		// Author image - only if image display is enabled in attributes.
-		if ( ! empty( $author['image'] ) && ( ! isset( $attributes['showImage'] ) || $attributes['showImage'] ) ) {
-			$html .= $this->render_author_image( $author );
-		}
-
-		$html .= '</div>'; // Close .apb-author-list-left.
-
-		// Right column with author details.
-		$html .= '<div class="apb-author-list-right">';
-
-		// Author info header.
-		$html .= '<div class="apb-author-list-header">';
-
-		// Author name.
-		if ( ! empty( $author['title'] ) ) {
-			$html .= $this->render_author_name( $author );
-		}
-
-		// Author position if available and display is enabled.
-		if ( ! empty( $author['position'] ) && ( ! isset( $attributes['showPosition'] ) || $attributes['showPosition'] ) ) {
-			$html .= $this->render_author_position( $author );
-		}
-
-		$html .= '</div>'; // Close .apb-author-list-header.
-
-		// Author email - only if email display is enabled in attributes.
-		if ( ! empty( $author['email'] ) && ( ! isset( $attributes['showEmail'] ) || $attributes['showEmail'] ) ) {
-			$html .= $this->render_author_email( $author );
-		}
-
-		// Author description - only if description display is enabled in attributes.
-		if ( ! empty( $author['description'] ) && ( ! isset( $attributes['showDescription'] ) || $attributes['showDescription'] ) ) {
-			$html .= $this->render_author_description( $author );
-		}
-
-		// Social profiles in footer if enabled.
-		if ( ! empty( $author['social'] ) && is_array( $author['social'] ) && ( ! isset( $attributes['showSocial'] ) || $attributes['showSocial'] ) ) {
-			$html .= '<div class="apb-author-list-footer">';
-			$html .= $this->render_social_profiles( $author['social'], 'apb-list-social' );
-			$html .= '</div>';
-		}
-
-		$html .= '</div>'; // Close .apb-author-list-right.
-		$html .= '</div>'; // Close .apb-author-list-detailed.
-
-		return $html;
-	}
-
-	/**
-	 * Render author image section.
-	 *
-	 * @param array  $author        Author data.
-	 * @param string $wrapper_class Optional. Additional CSS class for the image container.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_author_image( array $author, string $wrapper_class = '' ): string {
-		$classes = 'apb-author-image';
-		if ( ! empty( $wrapper_class ) ) {
-			$classes .= ' ' . $wrapper_class;
-		}
-
-		return '<div class="' . esc_attr( $classes ) . '">' .
-		       '<img src="' . esc_url( $author['image'] ) . '" alt="' .
-		       esc_attr( $author['title'] ) . '" loading="lazy" />' .
-		       '</div>';
-	}
-
-	/**
-	 * Render author name section.
-	 *
-	 * @param array $author Author data.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_author_name( array $author ): string {
-		return '<h3 class="apb-author-name">' . esc_html( $author['title'] ) . '</h3>';
-	}
-
-	/**
-	 * Render author position section.
-	 *
-	 * @param array $author Author data.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_author_position( array $author ): string {
-		return '<div class="apb-author-position">' . esc_html( $author['position'] ) . '</div>';
-	}
-
-	/**
-	 * Render author email section.
-	 *
-	 * @param array $author Author data.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_author_email( array $author ): string {
-		return '<div class="apb-author-email">' .
-		       '<a href="mailto:' . esc_attr( $author['email'] ) . '">' .
-		       esc_html( $author['email'] ) . '</a>' .
-		       '</div>';
-	}
-
-	/**
-	 * Render author description section.
-	 *
-	 * @param array $author Author data.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_author_description( array $author ): string {
-		return '<div class="apb-author-description">' .
-		       wp_kses_post( $author['description'] ) .
-		       '</div>';
-	}
-
-	/**
-	 * Render social profiles section.
-	 *
-	 * @param array  $profiles      Social profile URLs.
-	 * @param string $wrapper_class Optional. Additional CSS class for the social profiles container.
-	 *
-	 * @return string Rendered HTML.
-	 */
-	private function render_social_profiles( array $profiles, string $wrapper_class = '' ): string {
-		$classes = 'apb-social-profiles';
-		if ( ! empty( $wrapper_class ) ) {
-			$classes .= ' ' . $wrapper_class;
-		}
-
-		$html = '<div class="' . esc_attr( $classes ) . '">';
-		$html .= '<ul class="apb-social-list">';
-
-		$social_icons = array(
-			'facebook'  => 'dashicons-facebook',
-			'twitter'   => 'dashicons-twitter',
-			'linkedin'  => 'dashicons-linkedin',
-			'instagram' => 'dashicons-instagram',
-			'website'   => 'dashicons-admin-site',
-		);
-
-		foreach ( $profiles as $network => $url ) {
-			if ( ! empty( $url ) && isset( $social_icons[ $network ] ) ) {
-				$html .= '<li class="apb-social-item apb-social-' . esc_attr( $network ) . '">';
-				$html .= '<a href="' . esc_url( $url ) . '" target="_blank" rel="noopener noreferrer">';
-				$html .= '<span class="dashicons ' . esc_attr( $social_icons[ $network ] ) . '" aria-hidden="true"></span>';
-				$html .= '<span class="screen-reader-text">' . esc_html( ucfirst( $network ) ) . '</span>';
-				$html .= '</a>';
-				$html .= '</li>';
-			}
-		}
-
-		$html .= '</ul>';
-		$html .= '</div>';
 
 		return $html;
 	}
